Fix many warnings

+++ lib/gdt.c
@@ -16,7 +16,7 @@ void gdt_set_gate(int num, unsigned long base, unsigned long limit, unsigned cha
 void gdt_install()
 {
     gp.limit = (sizeof(struct gdt_entry) * 3) - 1;
-    gp.base = &gdt;
+    gp.base = (unsigned int)&gdt;
 
     gdt_set_gate(0, 0, 0, 0, 0);
